feat(keybroker-client): Send event log to the server when available
First, try to send the evidence on its own, as usual. If that doesn't
work and an event log is available in sysfs, send it to the server
wrapped in a CMW. The server may use it to reconstruct the reference
values.

In Mock mode, send an empty event log to test wrapping.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>