feat(keybroker-server): Parse event log to construct reference values
Use the event log library from cca-realm-measurements to recreate the
reference values, if they aren't provided. This is a proof of concept to
show how we can generate the RIM dynamically, and cache it for
subsequent attestations.

Some work is still needed to add policy tests on the log itself, and to
support the REMs (see comments).

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>